% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printVersion.R
\name{printVersion}
\alias{printVersion}
\title{Linear model}
\usage{
printVersion()
}
\arguments{
\item{formula}{formula to be fitted}

\item{data}{data}

\item{...}{other arguments}
}
\description{
This function estimates model parameters in a linear regression model
}
